[MOD] moved chat components into component folder and chatRoom page to user/shared folder, split components into different files, renamed some components
