--- src/Users/Messages/Messages.scss
@@ -1,153 +1,0 @@
-.messaging-page {
-  display: flex;
-  margin-top: 25px;
-  max-width: 800px;
-  margin: 0 auto;
-  background-color: #fff;
-  border-radius: 5px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.user-list {
-  flex-basis: 30%;
-  padding: 20px;
-  border-right: 1px solid #e1e8ed;
-}
-
-.message-list {
-  flex-grow: 1;
-  padding: 20px;
-}
-
-/* Styles pour la liste des utilisateurs */
-.user {
-  display: flex;
-  align-items: center;
-  padding: 10px;
-  border-radius: 5px;
-  cursor: pointer;
-}
-
-.user:hover {
-  background-color: #f5f8fa;
-}
-
-.user-avatar {
-  width: 40px;
-  height: 40px;
-  margin-right: 10px;
-  border-radius: 50%;
-  background-color: #4f23e2;
-}
-
-.user-username {
-  font-weight: bold;
-}
-
-.user-status {
-  font-size: 12px;
-  color: #8899a6;
-}
-
-/* Styles pour un message individuel */
-.message {
-  margin-bottom: 20px;
-  padding: 10px;
-  background-color: #f5f8fa;
-  border-radius: 5px;
-}
-
-.message-header {
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 5px;
-}
-
-.message-username {
-  font-weight: bold;
-}
-
-.message-time {
-  color: #8899a6;
-}
-
-.message-content {
-  margin-top: 5px;
-}
-
-/* Styles pour la zone de saisie et le bouton d'envoi */
-.message-input {
-  display: flex;
-  align-items: center;
-  padding: 10px;
-  border-top: 1px solid #4f23e2;
-}
-
-.message-input input[type="text"] {
-  flex-grow: 1;
-  padding: 8px;
-  border: none;
-  border-radius: 20px;
-  background-color: #f5f8fa;
-  outline: none;
-}
-
-.message-input label.file-input-label {
-  position: relative;
-}
-
-.message-input input[type="file"] {
-  position: absolute;
-  left: -9999px;
-}
-
-.message-input .file-input-button {
-  margin-left: 10px;
-  padding: 8px;
-  border-radius: 50%;
-  background-color: #4f23e2;
-  color: #fff;
-  font-weight: bold;
-  cursor: pointer;
-}
-
-.message-input .file-input-button:hover {
-  background-color: #3c1eaa;
-}
-
-.message-input .file-input-button:focus {
-  outline: none;
-}
-
-/* Styles pour les boutons */
-button.new-conversation-button,
-button.send-button {
-  margin-left: 10px;
-  padding: 8px 16px;
-  border: none;
-  border-radius: 20px;
-  background-color: #4f23e2;
-  color: #fff;
-  font-weight: bold;
-  cursor: pointer;
-}
-
-button.new-conversation-button:hover,
-button.send-button:hover {
-  background-color: #3c1eaa;
-}
-
-button.new-conversation-button:focus,
-button.send-button:focus {
-  outline: none;
-}
-
-/* Styles pour les liens */
-a {
-  color: #4f23e2;
-  text-decoration: none;
-}
-
-a:hover {
-  text-decoration: underline;
-}
