+++ src/index.jsx
@@ -11,12 +11,12 @@ import ForgottenPasswordPage from './Users/Public/forgottenPasswordPage'
 import NoPage from './Users/Public/noPage'
 import StudentHomePage from './Users/Student/dashboardStudent'
 import TeacherHomePage from './Users/Teacher/dashboardTeacher'
-import Messages from './Users/Messages/index'
 import NewFormPage from './Users/Teacher/newFormPage'
 import FormListStudentPage from './Users/Student/formListStudentPage'
 import FormStudentPage from './Users/Student/formStudentPage'
 import FormListTeacherPage from './Users/Teacher/formListTeacherPage'
 import FormTeacherPage from './Users/Teacher/formTeacherPage'
+import ChatRoomPage from './Users/Shared/chatRoomPage'
 
 const rootElement = document.getElementById('root')
 
@@ -36,21 +36,21 @@ if (rootElement) {
           <>
             <Route path='/login' element={<Navigate to='/' replace />} />
             <Route path='/forgot' element={<Navigate to='/' replace />} />
-            <Route path='/messages' element={<Messages />} />
+            <Route path='/messages' element={<ChatRoomPage />} />
           </>
         )}
         {sessionStorage.getItem('role') === 'admin' && (
           <>
             <Route path='/' element={<AdmHomePage />} />
             <Route path='/accounts' element={<AdmAccountsPage />} />
-            <Route path='/messages' element={<Messages />} />
+            <Route path='/messages' element={<ChatRoomPage />} />
           </>
         )}
         {sessionStorage.getItem('role') === 'administration' && (
           <>
             <Route path='/' element={<SchoolAdmHomePage />} />
             <Route path='/accounts' element={<SchoolAdmAccountsPage />} />
-            <Route path='/messages' element={<Messages />} />
+            <Route path='/messages' element={<ChatRoomPage />} />
           </>
         )}
         {sessionStorage.getItem('role') === 'student' && (
@@ -63,7 +63,7 @@ if (rootElement) {
         {sessionStorage.getItem('role') === 'teacher' && (
           <>
             <Route path='/' element={<TeacherHomePage />} />
-            <Route path='/messages' element={<Messages />} />
+            <Route path='/messages' element={<ChatRoomPage />} />
             <Route path='/questionnaires' element={<FormListTeacherPage />} />
             <Route path='/questionnaire' element={<NewFormPage />} />
             <Route path='/questionnaire/:id' element={<FormTeacherPage />} />
